requirements.txt
langchain>=0.3.15
langgraph>=0.3.15
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pydantic>=2.5.0
duckduckgo-search>=2025.4.1
langchain-community>=0.3.15
coinbase-agentkit==0.3.0
coinbase-agentkit-langchain==0.2.0
eth-account>=0.11.0
cdp-sdk==0.21.0
jsonschema>=4.23.0
nilql>=0.0.0a12
paramiko>=3.5.1

